tests/wpt/web-platform-tests/service-workers/service-worker/update.https.html

<!DOCTYPE html>
<title>Service Worker: Registration update()</title>
<meta name="timeout" content="long">
<script src="/common/utils.js"></script>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/testharness-helpers.js"></script>
<script src="resources/test-helpers.sub.js"></script>
<script>
'use strict';

const SCOPE = 'resources/simple.txt';
const WORKER_URL_BASE = 'resources/update-worker.py';

async function prepare_ready_registration(t, mode) {
  const key = token();
  const worker_url = `${WORKER_URL_BASE}?Key=${key}&Mode=${mode}`;
  const expected_url = normalizeURL(worker_url);
  const registration = await service_worker_unregister_and_register(
      t, worker_url, SCOPE);
  await wait_for_state(t, registration.installing, 'activated');
  assert_equals(registration.installing, null,
                'prepare_ready: installing');
  assert_equals(registration.waiting, null,
                'prepare_ready: waiting');
  assert_equals(registration.active.scriptURL, expected_url,
                'prepare_ready: active');
  return [registration, expected_url];
}

function assert_installing_and_active(registration, expected_url) {
  assert_equals(registration.installing.scriptURL, expected_url,
                'assert_installing_and_active: installing');
  assert_equals(registration.waiting, null,
                'assert_installing_and_active: waiting');
  assert_equals(registration.active.scriptURL, expected_url,
                'assert_installing_and_active: active');
}

function assert_waiting_and_active(registration, expected_url) {
  assert_equals(registration.installing, null,
                'assert_waiting_and_active: installing');
  assert_equals(registration.waiting.scriptURL, expected_url,
                'assert_waiting_and_active: waiting');
  assert_equals(registration.active.scriptURL, expected_url,
                'assert_waiting_and_active: active');
}

function assert_active_only(registration, expected_url) {
  assert_equals(registration.installing, null,
                'assert_active_only: installing');
  assert_equals(registration.waiting, null,
                'assert_active_only: waiting');
  assert_equals(registration.active.scriptURL, expected_url,
                'assert_active_only: active');
}

promise_test(async t => {
  const [registration, expected_url] =
      await prepare_ready_registration(t, 'normal');
  t.add_cleanup(() => registration.unregister());

  await Promise.all([registration.update(), wait_for_update(t, registration)]);
  assert_installing_and_active(registration, expected_url);

  await wait_for_state(t, registration.installing, 'installed');
  assert_waiting_and_active(registration, expected_url);

  await wait_for_state(t, registration.waiting, 'activated');
  assert_active_only(registration, expected_url);
}, 'update() should succeed when new script is available.');

promise_test(async t => {
  const [registration, expected_url] =
      await prepare_ready_registration(t, 'bad_mime_type');
  t.add_cleanup(() => registration.unregister());

  promise_rejects(t, 'SecurityError', registration.update());
  assert_active_only(registration, expected_url);
}, 'update() should fail when mime type is invalid.');

promise_test(async t => {
  const [registration, expected_url] =
      await prepare_ready_registration(t, 'redirect');
  t.add_cleanup(() => registration.unregister());

  promise_rejects(t, new TypeError(), registration.update())
  assert_active_only(registration, expected_url);
}, 'update() should fail when a response for the main script is redirect.');

promise_test(async t => {
  const [registration, expected_url] =
      await prepare_ready_registration(t, 'syntax_error');
  t.add_cleanup(() => registration.unregister());

  promise_rejects(t, new TypeError(), registration.update());
  assert_active_only(registration, expected_url);
}, 'update() should fail when a new script contains a syntax error.');

promise_test(async t => {
  const [registration, expected_url] =
      await prepare_ready_registration(t, 'throw_install');
  t.add_cleanup(() => registration.unregister());

  await Promise.all([registration.update(), wait_for_update(t, registration)]);
  assert_installing_and_active(registration, expected_url);
}, 'update() should resolve when the install event throws.');

promise_test(async t => {
  const [registration, expected_url] =
      await prepare_ready_registration(t, 'normal');
  t.add_cleanup(() => registration.unregister());

  // We need to hold a client alive so that unregister() below doesn't remove
  // the registration before update() has had a chance to look at the pending
  // uninstall flag.
  const frame = await with_iframe(SCOPE);
  t.add_cleanup(() => frame.remove());

  promise_rejects(
      t, new TypeError(),
      Promise.all([registration.unregister(), registration.update()]));
}, 'update() should fail when the pending uninstall flag is set.')
</script>
